Little refactoring

+++ SystemFiles.h
@@ -33,12 +33,16 @@ public:
 		int numberIndexBlock = disc->TrySaveFile(fileContent, realSize, occupiedSpace);
 		if (numberIndexBlock == -1)
 			return -1;
+		if (getFile(fileName))
+			return -2;
 		File* file = new File(fileName, numberIndexBlock, occupiedSpace, realSize, access);
 		files.push_back(file);
 		return 0;
 	}
 	int EditFile(string newFileName, string oldFileName, string fileContent)
 	{
+		if (!getFile(oldFileName))
+			return -2;
 		int realSize = 0;
 		int numberIndexBlock = checkNumberIndexBlock(oldFileName);
 		if (numberIndexBlock != -1)
@@ -102,6 +106,8 @@ public:
 	}
 	char* OpenFile(string fileName)
 	{
+		if (!getFile(fileName))
+			return "";
 		int numberSector = checkNumberIndexBlock(fileName);
 		if (numberSector > -1 && numberSector < amountSectors)
 			return disc->TryOpenFile(numberSector);
@@ -109,6 +115,8 @@ public:
 	}
 	void DeleteFile(string fileName)
 	{
+		if (!getFile(fileName))
+			return;
 		int numberSector = -1;
 		File* file = getFile(fileName);
 		disc->DeleteFile(file->GetNumberSector());
@@ -134,7 +142,10 @@ public:
 			return true;
 		else return false;
 	}
-
+	int GetSizeFreeMemory()
+	{
+		return disc->AmoutFreeSectors()*sizeSector;
+	}
 	private:
 		void saveDefaultData()
 		{
